% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{autoDimnames}
\alias{autoDimnames}
\title{Assign Dimnames with Sanity Checks}
\usage{
autoDimnames(x, value)
}
\arguments{
\item{x}{A matrix-like object with two dimensions.}

\item{value}{A list of length 2, containing the new row and column names
respectively.}
}
\value{
The modified object \code{x} with updated dimnames.
}
\description{
A helper function for safely assigning dimnames to a matrix-like object.
Primarily used in method dispatch (e.g., \verb{dimnames<-} for \code{FlexAssays}).
}
\details{
If the input \code{value} is not a list of length 2, will raise an error.
If the list is empty but the object \code{x} has non-zero rows or columns,
a warning is issued and the original object is returned unmodified to
prevents silent loss of dimension names on non-empty objects.

This function is intended to be used inside \code{setMethod("dimnames<-", ...)}
and not called directly by end-users.
}
\examples{
mat <- matrix(1:4, nrow = 2)
autoDimnames(mat, list(c("A", "B"), c("X", "Y")))

# Remove all dimension names
autoDimnames(mat, list(NULL, NULL))

}
