% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexAssays-development.R
\name{cleanDimMaps}
\alias{cleanDimMaps}
\title{Clean Unmapped Dimensions for FlexAssays}
\usage{
cleanDimMaps(x, check = TRUE, ...)
}
\arguments{
\item{x}{A \code{\link{FlexAssays}} object}

\item{check}{Logical. If \code{TRUE} (default), \code{\link[=validObject]{validObject()}} will be called
after cleaning to ensure the object remains valid.}

\item{...}{Arguments passed to other methods.}
}
\value{
A \code{\link{FlexAssays}} object with cleaned dimension maps.
}
\description{
Removes unmapped row and column names from the \code{rowMap} and \code{colMap} of a
\code{\link{FlexAssays}} object, effectively trimming unused dimensions. This
function is primarily intended for developers or pipeline maintainers to
optimize memory and ensure consistency in dimension maps.
}
\examples{
fa <- exampleFlexAssays()
fa[[2]] <- NULL

fa <- cleanDimMaps(fa)

}
