% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexAssays.R
\name{FlexAssays-assays}
\alias{FlexAssays-assays}
\alias{assays,SimpleFlexAssays-method}
\alias{assays<-,FlexAssays,list_OR_List-method}
\alias{assays<-,FlexAssays,NULL-method}
\alias{assay,FlexAssays,missing-method}
\alias{assay,FlexAssays,character-method}
\alias{assay,FlexAssays,numeric-method}
\alias{assay<-,FlexAssays,missing-method}
\alias{assay<-,FlexAssays,numeric-method}
\alias{assay<-,FlexAssays,character-method}
\title{Get or set assays}
\usage{
\S4method{assays}{SimpleFlexAssays}(x, withDimnames = TRUE, ...)

\S4method{assays}{FlexAssays,list_OR_List}(x, withDimnames = TRUE, ...) <- value

\S4method{assays}{FlexAssays,NULL}(x, withDimnames = TRUE, ...) <- value

\S4method{assay}{FlexAssays,missing}(x, i, withDimnames = TRUE, ...)

\S4method{assay}{FlexAssays,character}(x, i, withDimnames = TRUE, ...)

\S4method{assay}{FlexAssays,numeric}(x, i, withDimnames = TRUE, ...)

\S4method{assay}{FlexAssays,missing}(x, i, withDimnames = TRUE, ...) <- value

\S4method{assay}{FlexAssays,numeric}(x, i, withDimnames = TRUE, ...) <- value

\S4method{assay}{FlexAssays,character}(x, i, withDimnames = TRUE, ...) <- value
}
\arguments{
\item{x}{A \code{\link{FlexAssays}} object.}

\item{withDimnames}{Logical, indicating whether to restore the dimnames of
assays from \code{rowMap} or \code{colMap}. Default is \code{TRUE}.}

\item{...}{Arguments passed to other methods.}

\item{value}{A single matrix-like object to set as the \code{i}-th assay. Must
contain both row and column names. If \code{NULL}, the original assay at index \code{i}
will be removed.}

\item{i}{Index of the assay to access or modify.}
}
\value{
\itemize{
\item \code{assays()}: Returns a list of matrix-like assays, optionally with
restored dimension names.
}

\itemize{
\item \code{assay}: Returns the matrix-like assay object at the \code{i}-th
element.
}

\itemize{
\item \verb{assay<-}: Returns an updated \code{\link{FlexAssays}} with the
\code{i}-th assay replaced by \code{new.assay}.
}
}
\description{
Methods to access or update the assay list for a  \code{\link{FlexAssays}}.
}
\details{
\itemize{
\item \code{assays} method must be implemented for each concrete sub-class of
\code{FlexAssays}.
}

\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::assays<-}} method is defined for the VIRTUAL class \code{\link{FlexAssays}} and
should automatically ensure the output \code{x} is validated. Any sub-class that
implements \code{\link{setRawAssays}} and \code{\link{cleanRawAssays}} correctly should works for
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::assays<-}}.
}
\examples{
## assays
fa <- exampleFlexAssays()

assays(fa)

## assays<-
fa <- exampleFlexAssays()
aa <- assays(fa)[1:2]
assays(fa) <- aa

## Remove all assays
fa <- exampleFlexAssays()
assays(fa) <- NULL

## assay
fa <- exampleFlexAssays()
assay(fa, 2)

## assay<-
fa <- exampleFlexAssays()
a <- assay(fa, 1)
assay(fa, 4) <- a

}
\seealso{
\code{\link{setRawAssays}} and \code{\link{cleanRawAssays}}
}
