% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexAssays-development.R
\name{FlexAssays-flex}
\alias{FlexAssays-flex}
\alias{rowFlex}
\alias{rowFlex,FlexAssays-method}
\alias{colFlex}
\alias{colFlex,FlexAssays-method}
\alias{assayClasses}
\alias{assayClasses,FlexAssays-method}
\title{Control Flexibility for FlexAssays}
\usage{
rowFlex(x)

\S4method{rowFlex}{FlexAssays}(x)

colFlex(x)

\S4method{colFlex}{FlexAssays}(x)

assayClasses(x)

\S4method{assayClasses}{FlexAssays}(x)
}
\arguments{
\item{x}{A \code{\link{FlexAssays}} object}
}
\value{
\itemize{
\item \code{rowFlex}, \code{colFlex}: A character scalar indicating the rule of row or
column flexibility.
}

\itemize{
\item \code{assayClasses}: A character vector indicating the supported classes for
assay elements.
}
}
\description{
Methods for retrieving the configuration parameters that control row/column
flexibility and supported assay classes in a \code{\link{FlexAssays}} object.
}
\details{
The flexibility of rows and columns in \code{\link{FlexAssays}} is governed
by one of the following rules:
\itemize{
\item \code{free}: Global row/column names can be extended when adding assays with
new dimension names.
\item \code{bounded}: Global row/column names are fixed; assays with unmatched
names \emph{\strong{cannot}} be added.
\item \code{fixed}: All assays must share the exact same row/column names.
}
}
\examples{
fa <- exampleFlexAssays()
rowFlex(fa)
colFlex(fa)
assayClasses(fa)

}
