% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexAssays-development.R
\name{FlexAssays-map}
\alias{FlexAssays-map}
\alias{rowMap}
\alias{rowMap,FlexAssays-method}
\alias{colMap}
\alias{colMap,FlexAssays-method}
\title{Dimensional map for FlexAssays}
\usage{
rowMap(x)

\S4method{rowMap}{FlexAssays}(x)

colMap(x)

\S4method{colMap}{FlexAssays}(x)
}
\arguments{
\item{x}{A \code{\link{FlexAssays}} object}
}
\value{
A logical \code{\link[SparseArray]{SVT_SparseMatrix}} indicating the presence of
global row or column names across individual assays in \code{x}.
}
\description{
Accessors for the row/column mapping matrices of a \code{\link{FlexAssays}}
object. These logical maps indicate which rows or columns are present in each
assay. Primarily intended for developers rather than end-users.
}
\examples{
fa <- exampleFlexAssays()
rowMap(fa)
colMap(fa)

}
