% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexAssays.R
\name{FlexAssays-methods}
\alias{FlexAssays-methods}
\alias{show,FlexAssays-method}
\alias{length,FlexAssays-method}
\alias{names,FlexAssays-method}
\alias{names<-,FlexAssays-method}
\alias{dim,FlexAssays-method}
\alias{dimnames,FlexAssays-method}
\alias{dimnames<-,FlexAssays,ANY-method}
\alias{rownames<-,FlexAssays-method}
\alias{colnames<-,FlexAssays-method}
\alias{as.list.FlexAssays}
\title{Methods for FlexAssays}
\usage{
\S4method{show}{FlexAssays}(object)

\S4method{length}{FlexAssays}(x)

\S4method{names}{FlexAssays}(x)

\S4method{names}{FlexAssays}(x) <- value

\S4method{dim}{FlexAssays}(x)

\S4method{dimnames}{FlexAssays}(x)

\S4method{dimnames}{FlexAssays,ANY}(x) <- value

\S4method{rownames}{FlexAssays}(x) <- value

\S4method{colnames}{FlexAssays}(x) <- value

\method{as.list}{FlexAssays}(x, ...)
}
\arguments{
\item{x, object}{A \code{\link{FlexAssays}} object}

\item{value}{An object of a class specified in the S4 method signature. Typically a character vector for names or
dimension names.}

\item{...}{Arguments passed to other methods.}
}
\value{
\itemize{
\item \code{show}: Print the basic information of \code{object} to the console.
}

\itemize{
\item \code{length}: Returns an integer indicating the number of assays.
}

\itemize{
\item \code{names}: A character vector specifying the layer names.
}

\itemize{
\item \verb{names<-}: Returns \code{x} with updated assay names.
}

\itemize{
\item \code{dim}: Returns an integer vector of length 2 representing the global
dimensions (rows × columns).
}

\itemize{
\item \code{dimnames}: Returns a list of global row and column names.
}

\itemize{
\item \verb{dimnames<-}, \verb{rownames<-} and \verb{colnames<-}: Return an updated \code{x} with
modified row/column names.
}

\itemize{
\item \code{as.list}: Returns a list of assays extracted from \code{x}.
}
}
\description{
A collection of accessor methods for retrieving or modifying basic structural
information from a \code{\link{FlexAssays}} object.
}
\note{
These methods are primarily intended for interacting with the top-level
structure of a \code{\link{FlexAssays}} object.
}
\examples{
# Create some sparse matrices
fa <- exampleFlexAssays()
fa

length(fa)

names(fa)

names(fa) <- paste0("assay", as.character(seq_along(fa)))

dim(fa)
nrow(fa)
ncol(fa)

}
