% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{FlexAssays-package}
\alias{FlexAssays-package}
\title{FlexAssays: Flexible Container for Matrix-Like Assays with Logical Dimension Mapping}
\description{
A framework for organizing multiple matrix-like assays with different dimensions. Uses logical matrices to manage shared or unique row/column names across assays, while controlling dimension flexibility through rule-based settings. Designed for developers handling complex assay structures in bioinformatics workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bioconductor.org/packages/FlexAssays}
  \item Report bugs at \url{https://github.com/ycli1995/FlexAssays/issues}
}

}
\author{
\strong{Maintainer}: Yuchen Li \email{ycli1995@outlook.com} (\href{https://orcid.org/0009-0003-8612-4425}{ORCID})

}
\keyword{internal}
