% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexAssays-subset.R
\name{FlexAssays-subset}
\alias{FlexAssays-subset}
\alias{[[,FlexAssays,ANY,missing-method}
\alias{[[<-,FlexAssays,ANY,missing-method}
\alias{.DollarNames.FlexAssays}
\alias{$.FlexAssays}
\alias{$<-.FlexAssays}
\alias{[,FlexAssays,ANY,ANY,ANY-method}
\title{Subset a FlexAssays}
\usage{
\S4method{[[}{FlexAssays,ANY,missing}(x, i, j, ...)

\S4method{[[}{FlexAssays,ANY,missing}(x, i, j, ...) <- value

\method{.DollarNames}{FlexAssays}(x, pattern = "")

\method{$}{FlexAssays}(x, name)

\method{$}{FlexAssays}(x, name) <- value

\S4method{[}{FlexAssays,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{A \code{\link{FlexAssays}} object.}

\item{i, j}{Indices specifying elements to extract or replace. In \code{[[} or
\verb{[[<-}, argument \code{j} is ignored.}

\item{...}{Arguments passed to other methods.}

\item{value}{An object of a class specified in the S4 method signature. Typically a matrix-like object to be set. If is
\code{NULL}, the corresponding assay will be removed.}

\item{pattern}{A regular expression used in \code{.DollarNames} to filter names.
Only matching names are returned.}

\item{name}{The name of the assay to get or set using the \code{$} accessor.}

\item{drop}{Logical, whether or not to drop those empty layers after
subsetting.}
}
\value{
\itemize{
\item \code{[[}: Returns a matrix-like assay with dimension names for the \code{i}-th
element of \code{x}.
}

\itemize{
\item \verb{[[<-}: Returns an updated \code{FlexAssays} object with the \code{i}-th assay
modified.
}

\itemize{
\item \code{.DollarNames}: Returns a character vector of assay names matching
\code{pattern}, used for tab-completion.
}

\itemize{
\item \code{$}: Returns a matrix-like assay with the specified \code{name} in \code{x}.
}

\itemize{
\item \verb{$<-}: Returns an updated \code{FlexAssays} object with a new or modified
assay.
}

\itemize{
\item \code{[}: Returns a subset of the original \code{FlexAssays} object, optionally
dropping empty assays if \code{drop = TRUE}.
}
}
\description{
Provides methods for extracting or subsetting elements from a
\code{\link{FlexAssays}} object with indices or names.
}
\examples{
fa <- exampleFlexAssays()

fa[[1]]
fa[["m1"]]

fa[[4]] <- fa[[1]]

}
\seealso{
\code{\link[=FlexAssays]{FlexAssays()}} for object construction and structure.

\code{\link[=.DollarNames]{.DollarNames()}} for the S3 generic.
}
