% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexAssays-development.R
\name{FlexAssays-updateObject}
\alias{FlexAssays-updateObject}
\alias{updateObject}
\alias{updateObject,FlexAssays-method}
\title{Update a FlexAssays object}
\usage{
\S4method{updateObject}{FlexAssays}(object, ..., verbose = FALSE)
}
\arguments{
\item{object}{A legacy \code{\link{FlexAssays}} object that may need to be
updated.}

\item{...}{Arguments passed to other methods.}

\item{verbose}{Logical. If \code{TRUE}, messages about the update process will be
printed.}
}
\value{
An updated version of \code{object} compatible with the current \code{FlexAssays} class
structure.
}
\description{
Method to update \code{\link{FlexAssays}} objects to the latest version.
}
\examples{
fa <- exampleFlexAssays()
updateObject(fa)

}
