% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{identicalNoAttrs}
\alias{identicalNoAttrs}
\title{Compare Two Objects Ignoring Specific Attributes}
\usage{
identicalNoAttrs(x, y, ignore.attrs = NULL, ...)
}
\arguments{
\item{x}{First object to compare.}

\item{y}{Second object to compare.}

\item{ignore.attrs}{A character vector of attribute names to ignore during
comparison.}

\item{...}{Additional arguments passed to \code{\link{identical}}.}
}
\value{
A logical scalar. \code{TRUE} when \code{x} and \code{y} are identical (ignoring
the specified attributes).
}
\description{
This function compares two R objects using \code{\link{identical}}, but
allows users to ignore specific attributes during comparison. It is
especially useful when structural identity matters more than meta data.
}
\examples{
a <- matrix(1:4, 2)
b <- matrix(1:4, 2)
attr(a, "meta") <- "something"
identical(a, b)    # FALSE
identicalNoAttrs(a, b, "meta")   # TRUE

}
