% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical_map.R
\name{logmap-helpers}
\alias{logmap-helpers}
\alias{mappedRowNames<-}
\alias{mappedRowNames}
\alias{intersectedRows}
\alias{dropMapRows}
\alias{dropMapCols}
\alias{removeMapCols}
\alias{validLogMap}
\title{Helper Functions for Sparse Logical Maps}
\usage{
mappedRowNames(x, i, append = TRUE, ...) <- value

mappedRowNames(x, i, invert = FALSE, ...)

intersectedRows(x, type = c("all", "duplicated"), ...)

dropMapRows(x, ...)

dropMapCols(x, ...)

removeMapCols(x, i, ...)

validLogMap(x, immediate. = TRUE)
}
\arguments{
\item{x}{A sparse logical map, typically a \code{\link[SparseArray]{SVT_SparseMatrix}}}

\item{i}{Index or name of the column to access or modify in \code{x}.}

\item{append}{Logical. If \code{TRUE}, values not currently in \code{rownames(x)} will
be appended as new rows. If \code{FALSE}, adding non-existing values will raise
an error.}

\item{...}{Arguments passed to other methods.}

\item{value}{A new mapping to assign to column \code{i}. Must be one of the
followings:
\itemize{
\item A character vector: entries in \code{value} are set to \code{TRUE} for column
\code{i}.
\item \code{NULL}: all entries in column \code{i} are set to \code{FALSE}.
}}

\item{invert}{Logical. If \code{TRUE}, returns row names that are \strong{not} mapped
to column \code{i}.}

\item{type}{Type of row intersection to return:
\itemize{
\item \code{all}(the default): features shared by all columns.
\item \code{duplicated}: features shared by at least two columns.
}}

\item{immediate.}{Logical. If \code{TRUE} (default), validation errors are thrown
immediately. If \code{FALSE}, errors are collected silently and return. This can
be useful in context of \code{\link[methods:validObject]{validObject}}.}
}
\value{
\itemize{
\item \verb{mappedRowNames<-}: An updated logical map with modified mappings of
column \code{i}.
}

\itemize{
\item \code{mappedRowNames}: A character vector specifying row names of \code{x} mapped
to column \code{i}.
}

\itemize{
\item \code{intersectedRows}: An integer vector of row indices mapped to multiple
columns, depending on \code{type}.
}

\itemize{
\item \code{dropMapRows}: An updated map \code{x} with rows that are not mapped to any
column removed.
}

\itemize{
\item \code{dropMapCols}: An updated map \code{x} with columns that do not map to any
row removed.
}

\itemize{
\item \code{removeMapCols}: An updated \code{x} with column \code{i} removed.
}

\itemize{
\item \code{validLogMap}: Check whether \code{x} is a valid sparse logical map.
}
}
\description{
A set of utility functions for manipulating sparse logical maps, typically
used to track row mappings across observations. Includes methods for
accessing or modifying mappings, selecting intersected features, and dropping
unused rows or columns.
}
\examples{
# Create a logical map for known observations
obs <- letters[1:4]
map <- sparseLogMap(LETTERS[1:10], ncol = length(obs), names = obs)
map

# Set entries to a new column
mappedRowNames(map, 5) <- sample(LETTERS, 10)
mappedRowNames(map, "e") <- sample(LETTERS, 10)

# Set entries to an existing column
mappedRowNames(map, "c") <- sample(LETTERS, 10)

# Set 'c' to NULL. All values in column 'c' will be FALSE
mappedRowNames(map, "c") <- NULL
map

# Get row names that exist in a specific column
mappedRowNames(map, 1)
mappedRowNames(map, "e", invert = TRUE)

# Which rows contain names that are mapped to multiple columns.
intersectedRows(map)

# Drop rows whose names have no column mapped to.
dropMapRows(map)

# Drop columns that no row names are mapped to.
dropMapCols(map)

# Remove columns
removeMapCols(map, 1)
removeMapCols(map, c(1, 3))
removeMapCols(map, c("a", "e"))

}
