% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexAssays-development.R
\name{setRawAssays}
\alias{setRawAssays}
\alias{setRawAssays,SimpleFlexAssays,list_OR_List-method}
\alias{setRawAssays,SimpleFlexAssays,NULL-method}
\title{Update all assays with/without validation}
\usage{
setRawAssays(x, new.assays, check = TRUE, ...)

\S4method{setRawAssays}{SimpleFlexAssays,list_OR_List}(x, new.assays, check = TRUE, ...)

\S4method{setRawAssays}{SimpleFlexAssays,NULL}(x, new.assays, check = TRUE, ...)
}
\arguments{
\item{x}{A \code{\link{FlexAssays}} object.}

\item{new.assays}{A list of new matrix-like assays to replace the existing
ones in \code{x}. If is \code{NULL}, should remove all internal assays.}

\item{check}{Logical. If \code{TRUE} (default), \code{\link[=validObject]{validObject()}} will be called to
validate the object after replacement. Setting it to \code{FALSE} will skip checks
for \code{rowFlex}, \code{colFlex} and \code{assayClasses}.}

\item{...}{Arguments passed to other methods.}
}
\value{
Returns an updated \code{\link{FlexAssays}} object with \code{new.assays} set into the
corresponding slot.
}
\description{
This is intended for low-level manipulation, e.g., programmatically replacing
all matrix-like objects in a sub-class of \code{\link{FlexAssays}}.
}
\details{
This method is lower-level than \code{\link{assays<-}} and is only intended for
development usage. It can bypass the time consuming validation, therefore the
caller must ensure \code{new.assays} are fitted with the original \code{FlexAssays}.
}
