% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{showS4Title}
\alias{showS4Title}
\title{Display the S4 class title for an object}
\usage{
showS4Title(object, ...)
}
\arguments{
\item{object}{An S4 object.}

\item{...}{Additional strings to include in the output.}
}
\value{
No return value. This function is called for printing.
}
\description{
This is a helper function to print the class name of an S4 object in
a standard format. It is usually called within \code{\link{show}} methods.
}
\examples{
setClass("NewClass", slots = c(data = "character"))
showS4Title(new("NewClass"), "with character")

}
