% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subsetMatByDimNames}
\alias{subsetMatByDimNames}
\title{Subset a Matrix by Row and Column Names}
\usage{
subsetMatByDimNames(mat, rownames, colnames, drop = TRUE)
}
\arguments{
\item{mat}{A matrix-like object with row and column names.}

\item{rownames}{A character vector of row names to subset. Use zero-length
character to remove all rows.}

\item{colnames}{A character vector of column names to subset. Use zero-length
character to remove all columns.}

\item{drop}{Logical. Whether or not to drop dimensions (passed to matrix
subsetting).}
}
\value{
A subset of \code{mat} restricted to the provided \code{rownames} and
\code{colnames}.
}
\description{
Efficiently subsets a matrix or matrix-like object by provided row and column
names. If the input names are already in the same order as the matrix, the
subsetting will be ignored to improve performance.
}
\details{
If the input row/column names exactly match those in \code{mat}, the function
avoids lookup and directly returns the original matrix or slices it
efficiently.
}
\examples{
mat <- matrix(1:9, nrow = 3)
rownames(mat) <- c("geneA", "geneB", "geneC")
colnames(mat) <- c("cell1", "cell2", "cell3")

subsetMatByDimNames(mat, rownames = "geneA", colnames = c("cell2", "cell3"))
# Returns a 1x2 matrix with values from geneA and specified columns

}
