% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexAssays-class.R
\name{validFlexAssays}
\alias{validFlexAssays}
\title{Validation for FlexAssays}
\usage{
validFlexAssays(x, immediate. = TRUE)
}
\arguments{
\item{x}{A \code{\link{FlexAssays}} to check.}

\item{immediate.}{Logical indicating if the errors should be raised
immediately. Set \code{immediate. = FALSE} can be useful for context of
\code{\link[methods:validObject]{validObject}}.}
}
\value{
Returns an invisible \code{NULL} if \code{x} is a valid \code{\link{FlexAssays}}. If
\code{immediate. = TRUE}, this function will raise an error when \code{x} is invalid.
Otherwise, it will return a character vector that collects all errors.
}
\description{
Check whether a \code{\link{FlexAssays}} is valid.
}
\examples{
m1 <- matrix(1:10, 2, 5)
rownames(m1) <- letters[1:2]
colnames(m1) <- LETTERS[1:5]

fa <- FlexAssays(m1)
validFlexAssays(fa)

}
\seealso{
\code{\link[=getErrors]{getErrors()}}
}
