% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexAssays-class.R
\name{exampleFlexAssays}
\alias{exampleFlexAssays}
\title{A Mini Example for FlexAssays}
\usage{
exampleFlexAssays()
}
\value{
A \code{\link{FlexAssays}} object with 3 assays.
}
\description{
Generate a tiny \code{\link{FlexAssays}} object as example.
}
\examples{
exampleFlexAssays()

}
