% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexAssays-development.R
\name{FlexAssays-assays}
\alias{FlexAssays-assays}
\alias{assays,SimpleFlexAssays-method}
\alias{assay,FlexAssays,missing-method}
\alias{assay,FlexAssays,character-method}
\alias{assay,FlexAssays,numeric-method}
\alias{setRawAssays}
\alias{setRawAssays,SimpleFlexAssays,list_OR_List-method}
\alias{assay<-,FlexAssays,missing-method}
\alias{assay<-,FlexAssays,numeric-method}
\alias{assay<-,FlexAssays,character-method}
\title{Get or set assays without dimension names}
\usage{
\S4method{assays}{SimpleFlexAssays}(x, withDimnames = TRUE, ...)

\S4method{assay}{FlexAssays,missing}(x, i, withDimnames = TRUE, ...)

\S4method{assay}{FlexAssays,character}(x, i, withDimnames = TRUE, ...)

\S4method{assay}{FlexAssays,numeric}(x, i, withDimnames = TRUE, ...)

setRawAssays(x, new.assays, ..., check = TRUE)

\S4method{setRawAssays}{SimpleFlexAssays,list_OR_List}(x, new.assays, ..., check = TRUE)

\S4method{assay}{FlexAssays,missing}(x, i, withDimnames = TRUE, ...) <- value

\S4method{assay}{FlexAssays,numeric}(x, i, withDimnames = TRUE, ...) <- value

\S4method{assay}{FlexAssays,character}(x, i, withDimnames = TRUE, ...) <- value
}
\arguments{
\item{x}{A \code{\link{FlexAssays}} object.}

\item{withDimnames}{Logical, indicating whether to restore the dimnames of
assays from \code{rowMap} or \code{colMap}. Default is \code{TRUE}.}

\item{...}{Arguments passed to other methods.}

\item{i}{Index of the assay to access or modify.}

\item{new.assays}{A list of new matrix-like assays to replace the existing
ones in \code{x}.}

\item{check}{Logical. If TRUE (default), \code{\link[=validObject]{validObject()}} will be called to
validate the object after replacement. Setting it to \code{FALSE} will skip checks
for \code{rowFlex}, \code{colFlex} and \code{assayClasses}.}

\item{value}{A single matrix-like object to set as the \code{i}-th assay. Must
contain both row and column names. If \code{NULL}, the original assay at index \code{i}
will be removed.}
}
\value{
\itemize{
\item \code{assays()}: Returns a list of matrix-like assays, optionally with
restored dimension names.
}

\itemize{
\item \code{assay}: Returns the matrix-like assay object at the \code{i}-th
element.
}

\itemize{
\item \code{setRawAssays()}: Returns an updated \code{\link{FlexAssays}} object
with \code{new.assays} set into the corresponding slot.
}

\itemize{
\item \verb{assay<-}: Returns an updated \code{\link{FlexAssays}} with the
\code{i}-th assay replaced by \code{new.assay}.
}
}
\description{
Internal helper methods to access or update the assay list of a given
\code{\link{FlexAssays}}. These methods are intended for internal use during
development; end-users should not call these methods directly.
}
\details{
These low-level methods are designed for internal manipulation of the assay
storage slot. Any subclass of \code{\link{FlexAssays}} that implements S4
methods for \code{assays} and \code{setRawAssays} should work well with the common
methods in \code{\link{FlexAssays-methods}}
}
\examples{
## assays
fa <- exampleFlexAssays()
assays(fa)

## assay
fa <- exampleFlexAssays()
assay(fa, 2)

## setRawAssays
fa <- exampleFlexAssays()
assays <- assays(fa)
fa <- setRawAssays(fa, assays)

## assay<-
fa <- exampleFlexAssays()
a <- assay(fa, 1)
assay(fa, 4) <- a

}
\keyword{internal}
