% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexAssays-class.R
\docType{class}
\name{FlexAssays}
\alias{FlexAssays}
\alias{SimpleFlexAssays-class}
\alias{FlexAssays-class}
\title{The FlexAssays class}
\usage{
FlexAssays(
  assays = SimpleList(),
  rownames = NULL,
  colnames = NULL,
  assayClasses = NULL,
  rowFlex = "free",
  colFlex = "free",
  ...
)
}
\arguments{
\item{assays}{One of the following:
\itemize{
\item A single matrix-like object
\item A list of multiple matrix-like objects.
\item An existing \code{FlexAssays}.
}}

\item{rownames, colnames}{A character vector containing the expected row or
column names. Default is \code{NULL}, which will fetch the union of row/column
names in \code{assays}. Only used when \code{assays} is empty or an existing
\code{FlexAssays}.}

\item{assayClasses}{A character vector indicating the supported classes of
assays. Default is \code{NULL} indicating no restrict of assay class.}

\item{rowFlex, colFlex}{A character scalar indicating the flexibility of rows
or columns. Must be one of the following:
\itemize{
\item \code{free}: The global rows/columns can be extended when a new assay with
non-existing dimension names is added.
\item \code{bounded}: The global rows/columns \strong{cannot} be extended when a new
assay with non-existing dimension names is added.
\item \code{fixed}: All assays must contain the same row/column names.
}}

\item{...}{Arguments passed to other methods.}
}
\description{
The \code{FlexAssays} virtual class provide a formal abstraction of layered
matrix-like objects (\code{assays}) with potential common names of dimensions.
Rows and columns of each assay are mapped using sparse logical maps (\code{rowMap}
and \code{colMap}), so that users can track whether an entry exists in rows or
columns for a specific assay.
}
\details{
For a given \code{FlexAssays}, \code{rowFlex}, \code{colFlex} and \code{assayClasses} must only
be set when the instance is initialized. The end-users should not modify
these slots.
}
\section{Slots}{

\describe{
\item{\code{rowMap,colMap}}{An \code{\link[SparseArray]{SVT_SparseMatrix}} to map row (column)
names for each assay.}

\item{\code{rowFlex,colFlex}}{See the description for parameters in the constructor
function.}

\item{\code{assayClasses}}{See the description for parameters in the constructor
function.}

\item{\code{data}}{In \code{SimpleFlexAssays}, \code{data} represents a list storing all
matrix-like assays. Row names of each assay must exist in row names of
\code{rowMap}, and so do column names for \code{colMap}.}
}}

\examples{
# Create some sparse matrices
mat1 <- Matrix::rsparsematrix(nrow = 10, ncol = 15, density = 0.75)
dimnames(mat1) <- list(sample(LETTERS, 10), sample(letters, 15))

mat2 <- Matrix::rsparsematrix(nrow = 15, ncol = 12, density = 0.75)
dimnames(mat2) <- list(sample(LETTERS, 15), sample(letters, 12))

mat3 <- Matrix::rsparsematrix(nrow = 25, ncol = 10, density = 0.75)
dimnames(mat3) <- list(sample(LETTERS, 25), sample(letters, 10))

# Create a FlexAssays from a list
mats <- FlexAssays(list(m1 = mat1, m2 = mat2, m3 = mat3))
mats

# Create a FlexAssays from a single matrix
mats <- FlexAssays(mat1)
mats

# Create a FlexAssays from an existing FlexAssays
mats2 <- FlexAssays(
  mats,
  rownames = head(rownames(mats)),
  colnames = head(colnames(mats))
)

}
\seealso{
\code{\link[=rowFlex]{rowFlex()}} and \code{\link[=colFlex]{colFlex()}} to learn how to control the flexibility of
dimensions.
}
