% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resetDimNames}
\alias{resetDimNames}
\title{Reset Row and Column Names of a Matrix-like Object}
\usage{
resetDimNames(mat, rownames = NULL, colnames = NULL)
}
\arguments{
\item{mat}{A matrix-like object. Must support \code{\link{rownames}} and
\code{\link{colnames}} methods.}

\item{rownames}{A character vector for new row names, or \code{NULL} to remove
row names.}

\item{colnames}{A character vector for new column names, or \code{NULL} to remove
column names.}
}
\value{
The input matrix with updated row and/or column names. The default
arguments will clear original row names and column names.
}
\description{
Replaces the row and/or column names of a matrix or matrix-like object.
}
\examples{
mat <- matrix(1:4, nrow = 2)
resetDimNames(mat, rownames = c("g1", "g2"), colnames = c("c1", "c2"))
resetDimNames(mat) # returns mat with no row/column names
}
